feat(ui-refactor, polish-test): intro screen refactor
